requirements.txt
python-dotenv
requests
pydantic>=2.0
ShopifyAPI
pytest
